feat: Add script to generate csv database from refractiveindex.info
This commit adds a new script, generate_csv_database.py, which scrapes the refractiveindex.info database and retrieves material data in yaml files. The script then records the relevant data in a csv file, which is used in Optiland to retrieve material data. The script also includes functionality to read each file and record wavelength ranges. This script will be useful for maintaining an up-to-date database of materials.