refactor: Update Tolerancing class to accept method and tol arguments
The Tolerancing class in core.py has been updated to accept two new arguments: method and tol. The method argument specifies the optimization method to use, with a default value of 'generic'. The tol argument specifies the tolerance for the compensator optimizer, with a default value of 1e-5. This change allows for more flexibility in the optimization process.