refactor: Update Tolerancing class to use BaseSampler for perturbation
The Tolerancing class in the core module has been updated to use the BaseSampler class from the perturbation module for generating perturbation values. This change improves the flexibility and extensibility of the perturbation functionality. The add_perturbation method now takes a sampler object instead of a perturbation object. The Perturbation class is still used internally to create the perturbation objects. This commit also includes some minor code formatting changes.