feat: Add Mirror3D class for visualizing 3D mirror surfaces
Add a new class, Mirror3D, to represent a 3D mirror surface in the optiland.visualization module. This class inherits from the Surface3D class and provides methods for configuring the material properties of the mirror surface.

The Mirror3D class takes a mirror surface and an extent as arguments in its constructor. The _configure_material method is used to set the color, ambient, diffuse, specular, and specular power properties of the mirror surface actor.

This commit adds the Mirror3D class to the optiland.visualization.mirror module.