Wavefront Class modification (#127)
* Wavefront class modified to save pupil data

The _generate_field_data function was modified to return additional parameters such as the x,y,z coordinates of the pupil and the reference sphere radius.

* Update wavefront.py

The wavefront class was modified to access the x,y,z coordinates of the pupil and the reference sphere radius by calling on the data object.

* Update wavefront.py

The _generate_field_data method was merged into the _generate_data method. The _get_path_length method has been removed since it is redundant with the new structure.

The calls to _opd_image_to_xp could only be partially reduced, since tracing the chief ray returns a single array element while the opd array for the other rays has another dimension.

* Changes to _correct_tilt method

The _correct_tilt method has been modified to take the opds of the reference sphere and the wavefront in a list as arguments. This aims to reduce the number of calls to the _correct_tilt method in the _generate_data method.

* feat: add WavefrontData dataclass

* Refactor WavefrontData and Wavefront classes for clarity and consistency

- Renamed 'opd_map' to 'opd' in WavefrontData for better clarity.
- Updated docstrings in Wavefront class to improve readability and consistency.
- Simplified field, wavelength, and distribution resolution methods.
- Enhanced data computation logic in _compute_all method.

* Chore: Refactor distribution handling and data retrieval in Wavefront class

* refactor: Refactor Wavefront class for improved clarity and data handling

* fix: remove duplicate call to _opd_image_to_xp

* refactor: Update test_generate_data to validate new data structure and types

* fix: correct typos in test function names and update path length retrieval

* fix: update PSF computation to use wavefront data in new format

* fix: update loop in RMS wavefront error calculation and correct class name typo in tests

* fix: update delta calculation to use new wavefront data retrieval method

---------

Co-authored-by: Kramer <kdanielharrison@gmail.com>