Real world projects (#177)
* Create Cooke_Triplet_with_Stock_Lenses.ipynb

A notebook explaining how to build a Cooke-Triplet system using off-the-shelf lenses has been written. This could serve as a good starting point for possible DIY projects using Optiland.

* chore: move notebook to new gallery folder

* docs: introduce Real-World Projects section to the gallery

* docs: fix links to thor labs and edmund optics

* Create Scan_Lens_System_for_UV.ipynb

A notebook describing the design for a scan lens with a focus on applications in UV has been written. In principle the same design could also be extended for use in visible light.

* Update Scan_Lens_System_for_UV.ipynb

Minor corrections to the explanation

* Move notebook to real_world_projects

Hi Kramer. As you mentioned, copied it to docs/gallery\real_world_projects and created the reference in the rst file as well.

---------

Co-authored-by: Kramer <kdanielharrison@gmail.com>