refractive index computed for a given temperature (#203)
* Update material_file.py

* Update material_file.py

Updated to compute refractive index also as a function of temperature

* Update material_file.py

fixed 1 bug
sometimes the reference temperature exist in the yaml file but not the thermal dispersion coefficients (e.g. fused_silica). In that case there was a problem in the previous version

* chore: minor formatting changes

* docs: use Google-style doc strings

* refactor: add _apply_environmental_correction method, keeping mathematical implementation consistent with original

* chore: ensure environmental corrections are applied only when thermodynamic coefficients are available

* Update material_file.py

A correction need to be applied for the input wavelength to compute the base_relative_n in case of arbitrary temperature and pressure

* chore: run formatter

* tests: add structural tests for material thermal index calculations

* docs: add mattemilio to the list of contributors

---------

Co-authored-by: Kramer <kdanielharrison@gmail.com>