Return array for Abbe material extinction coefficient (#388)
* Return array for Abbe material extinction coefficient

Returning a float breaks ray tracing, as that relies on `be.any` which requires arrays / tensors

* Use zeros_like to return an array with the same shape as wavelength

* Add ray trace tests for materials

* Define materials on NavarroWideAngleEye initialization

When defining the materials on class initialization, they are always defined with the NumPy backend. This causes issues when switching to the torch backend, which in many places does not support NumPy arrays. Defining the materials inside the __init__ method ensures that they are created with the correct backend.